+++ docs/history.rst
@@ -4,6 +4,7 @@ Version History
 :compare:`Next <3.0.4...master>`
 --------------------------------
 - Add a transcoder for `application/x-www-formurlencoded`_
+- Add support for encoding :class:`decimal.Decimal`
 - Add type annotations (see :ref:`type-info`)
 - Return a "406 Not Acceptable" if the :http:header:`Accept` header values cannot be matched
   and there is no default content type configured